requirements/base.txt
Django==2.2.10
django-filter==2.2.0
django-environ==0.4.5
djangorestframework==3.11.0
django-extensions==2.2.8
drf-yasg==1.17.0
Markdown==3.2
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
Scrapy==1.8.0
pygal==2.4.0
pendulum==2.0.5

django-jsonfield==1.4.0
